@@ -19,8 +19,17 @@
         "resolve-url-loader": "^3.1.2",
         "sass": "^1.32.11",
         "sass-loader": "^11.0.1",
-        "vue": "^2.6.12",
-        "vue-loader": "^15.9.8",
+        "vue": "^3.2.31",
+        "vue-loader": "^16.8.3",
         "vue-template-compiler": "^2.6.12"
+    },
+    "dependencies": {
+        "@fortawesome/fontawesome-svg-core": "^1.3.0",
+        "@fortawesome/free-regular-svg-icons": "^6.1.1",
+        "@fortawesome/free-solid-svg-icons": "^6.0.0",
+        "@fortawesome/vue-fontawesome": "^3.0.0-5",
+        "gsap": "^3.10.2",
+        "vue-router": "^4.0.14"
     }
+    
 }
