@@ -18,18 +18,19 @@ class LoginController extends Controller
     public function login(Request $request)
     {
         if (!auth()->attempt($request->only('email', 'password'))) {
-                return response()->json([
-                    'message' => 'Invalid login details'
-                ], 401);
-            }
+            return response()->json([
+                'message' => 'Invalid login details'
+            ], 401);
+        }
 
         $user = $this->user->where('email', $request['email'])->firstOrFail();
-
         $token = $user->createToken('auth_token')->plainTextToken;
 
         return response()->json([
             'access_token' => $token,
+            'user' => $user,
             'token_type' => 'Bearer',
+            'success' => true,
         ]);
     }
 
