@@ -26,20 +26,35 @@ class RegisterController extends Controller
             'email'         => 'required|string|email|max:255|unique:users',
             'password'      => 'required|string|min:8',
         ]);
-            $user = $this->user->create([
-                'first_name'=> $validatedData['first_name'],
-                'last_name' => $validatedData['last_name'],
-                'birthday'  => $validatedData['birthday'],
-                'email'     => $validatedData['email'],
-                'password'  => Hash::make($validatedData['password']),
-            ]);
+        $user = $this->user->create([
+            'first_name'=> $validatedData['first_name'],
+            'last_name' => $validatedData['last_name'],
+            'birthday'  => $validatedData['birthday'],
+            'email'     => $validatedData['email'],
+            'password'  => Hash::make($validatedData['password']),
+        ]);
 
         $token = $user->createToken('auth_token')->plainTextToken;
 
         return response()->json([
             'access_token' => $token,
             'token_type' => 'Bearer',
+            "message" => true,
         ]);
+    }
+    public function update()
+    {
+        $user = $this->user->findOrfail($this->request->id)->first();
+        $user->update($this->request->all());
+        return response()->json([
+            'user' => $user,
+            "message" => true,
+        ]);
+    }
+
+    public function fetchUsers(){
+        $users = $this->user->all();
+        return $users;
     }
 
 }
